requirements.txt
torch>=2.1.0
transformers>=4.38.0
datasets>=2.14,<3
peft>=0.7.0
trl>=0.7.10
accelerate>=0.27.0
bitsandbytes>=0.41.0
sentence-transformers>=2.5.0
chromadb>=0.4.22
langchain>=0.1.0
langchain-community>=0.0.10
# Dependencias API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-multipart
numexpr
tavily-python
tqdm
python-dotenv